+++ all/promise_timeout/index.d.ts
@@ -2,4 +2,4 @@ export declare function promise_timeout<O>(
 	promise:(()=>Promise<O>)|Promise<O>,
 	ms:number,
 	error?:Error
-):Promise<O>
+):Promise<O>&{ cancel:(val:O)=>void }