+++ all/waitfor/index.d.ts
@@ -2,4 +2,4 @@ export declare function waitfor<V>(
 	fn:()=>Promise<V>|V,
 	timeout:number,
 	period?:number
-):Promise<V>&{ cancel: ()=>Promise<V> }
+):Promise<V>&{ cancel:()=>Promise<V> }